CASS-49 | Move exported pojos to api package
