+++ src/main/java/org/mule/modules/cassandradb/api/DataCenter.java
@@ -1,7 +1,7 @@
 /**
  * (c) 2003-2017 MuleSoft, Inc. The software in this package is published under the terms of the Commercial Free Software license V.1 a copy of which has been included with this distribution in the LICENSE.md file.
  */
-package org.mule.modules.cassandradb.internal.operation.parameters;
+package org.mule.modules.cassandradb.api;
 
 
 public class DataCenter {
