+++ src/main/java/org/mule/modules/cassandradb/internal/service/CassandraService.java
@@ -1,7 +1,7 @@
 package org.mule.modules.cassandradb.internal.service;
 
 import org.mule.connectors.commons.template.service.ConnectorService;
-import org.mule.modules.cassandradb.internal.operation.parameters.CreateKeyspaceInput;
+import org.mule.modules.cassandradb.api.CreateKeyspaceInput;
 import org.mule.runtime.extension.api.annotation.param.Content;
 
 public interface CassandraService extends ConnectorService {
