+++ src/main/java/org/mule/modules/cassandradb/internal/util/ReplicationStrategy.java
@@ -4,7 +4,7 @@
 package org.mule.modules.cassandradb.internal.util;
 
 import org.apache.commons.lang3.StringUtils;
-import org.mule.modules.cassandradb.internal.operation.parameters.CreateKeyspaceInput;
+import org.mule.modules.cassandradb.api.CreateKeyspaceInput;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
