+++ src/main/java/org/mule/modules/cassandradb/internal/util/builders/HelperStatements.java
@@ -6,7 +6,7 @@ package org.mule.modules.cassandradb.internal.util.builders;
 import com.datastax.driver.core.schemabuilder.CreateKeyspace;
 import com.datastax.driver.core.schemabuilder.SchemaBuilder;
 import com.datastax.driver.core.schemabuilder.SchemaStatement;
-import org.mule.modules.cassandradb.internal.operation.parameters.CreateKeyspaceInput;
+import org.mule.modules.cassandradb.api.CreateKeyspaceInput;
 import org.mule.modules.cassandradb.internal.util.ReplicationStrategy;
